+++ test/ActivityWriteHandlerSpec.hs
@@ -522,6 +522,7 @@ milkActivity =
                 , techRole = ReferenceProduct
                 , techActivityLinkId = supplierActId
                 , techProcessLinkId = Nothing
+                , techSupplierActivity = Nothing
                 , techLocation = ""
                 , techComment = Nothing
                 , techPedigree = Nothing