+++ test/AllocationSpec.hs
@@ -409,6 +409,7 @@ tech fid amount role =
         , techRole = role
         , techActivityLinkId = UUID.nil
         , techProcessLinkId = Nothing
+        , techSupplierActivity = Nothing
         , techLocation = ""
         , techComment = Nothing
         , techPedigree = Nothing