+++ test/PersistenceSpec.hs
@@ -429,6 +429,7 @@ milkActivity name =
                 , techRole = ReferenceProduct
                 , techActivityLinkId = supplierActId
                 , techProcessLinkId = Nothing
+                , techSupplierActivity = Nothing
                 , techLocation = ""
                 , techComment = Nothing
                 , techPedigree = Nothing